Make App testable
